Abstract comparison regarding persistant names out of _graphs.
Two graphs are not equivalent unless their persistant names match up.

It doesn't make sense to have a more general notion of equivalence, and it
is probably more expensive to compute such a thing, as well.

Recall that even if persistant names are identified, the number of
isomorphisms from even a small graph can be really large.

Signed-off-by: mulhern <amulhern@redhat.com>